% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapterremoval.R
\name{adapterremoval_usage}
\alias{adapterremoval_usage}
\title{Print available arguments for adapterremoval}
\usage{
adapterremoval_usage()
}
\value{
AdapterRemoval available arguments and their usage.
}
\description{
Print available arguments for adapterremoval.
Note that some arguments to the
adapterremoval are invalid if they are
already handled as explicit function arguments.
}
\examples{
adapterremoval_usage()

}
\references{
Schubert, Lindgreen, and Orlando (2016).
AdapterRemoval v2: rapid adapter trimming, identification, and read merging.
BMC Research Notes, 12;9(1):88.
}
\author{
Zheng Wei
}
