% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2-build}
\alias{bowtie2-build}
\alias{bowtie2_build}
\title{Interface to bowtie2-build of bowtie2-2.2.3}
\usage{
bowtie2_build(references, bt2Index, ..., overwrite = FALSE)
}
\arguments{
\item{references}{\code{Character} vector. The path to the files containing the references for which to
build a bowtie index.}

\item{bt2Index}{\code{Character} scalar. Write bowtie2 index data to files with this prefix: 'dir/basename'.
If the files with path like 'dir/basename.*.bt2' already exists, the function function will cast an error,
unless argument overwrite is \code{TRUE}.}

\item{...}{Additional arguments to be passed on to the binaries. See below for details.}

\item{overwrite}{\code{Logical}. Force overwriting of existing files if setting \code{TRUE}.}
}
\description{
This function can be use to call \code{bowtie2-build} that wrapped in shared library.
}
\details{
All additional arguments in ... are interpreted as additional parameters to be passed on to
bowtie2_build. All of them should be \code{Character} or \code{Numeric} scalar. You can put all aditional
arguments in one \code{Character}(e.g. "--threads 8 --quiet") with white space splited just like command line,
or put them in different \code{Character}(e.g. "--threads","8","--quiet").Note that some arguments to the
bowtie2_build will be ignored if they are already handled as explicit function arguments. See the output of
\code{bowtie2_build_usage()} for details about available parameters.
}
\examples{
# Create a new graph and add a cycle of nodes to it

# Get node information from this graph

#>   id type label deg inde
}
\references{
Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
