% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2_usage}
\alias{bowtie2_usage}
\title{Print available arguments for bowtie2}
\usage{
bowtie2_usage(indexType = "s")
}
\arguments{
\item{indexType}{\code{Character} scalar. Defaults to 's' which shows help
document of the small bowtie2-align binary. 
Can optionally pass 'l' which shows help document of the large 
bowtie2-align binary.}
}
\value{
bowtie2 available arguments and their usage.
}
\description{
Note that some arguments to the
bowtie2 are invalid if they are
already handled as explicit function arguments.
}
\examples{
\donttest{
bowtie2_usage()
bowtie2_usage(indexType = 'l')
}
}
\references{
Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read
alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
\author{
Zheng Wei, Rahul Varki
}
