% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2_version}
\alias{bowtie2_version}
\title{Print version information of bowtie2-2.2.3}
\usage{
bowtie2_version()
}
\references{
Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
