% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUtils.R
\name{checkFileCreateable}
\alias{checkFileCreateable}
\alias{checkFileCreatable}
\title{Check if file is creatable}
\usage{
checkFileCreatable(filePath, argname, overwrite)
}
\arguments{
\item{filePath}{\code{Character} scalar. Path to file to be created.}

\item{argname}{\code{Character} scalar. Name of the argument passed.}

\item{overwrite}{\code{Logical} Will warn user of overwriting if set \code{TRUE}.}
}
\value{
Indicates whether the file can be created at the path given.
}
\description{
This is an internal function that is not meant to be used outside
of the package. It determines whether the file at the end of the path can
be created.
}
\author{
Zheng Wei
}
