% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUtils.R
\name{checkFileExist}
\alias{checkFileExist}
\title{Check if file exists}
\usage{
checkFileExist(filePath, argname)
}
\arguments{
\item{filePath}{\code{Character} scalar. Path to file of interest.}

\item{argname}{\code{Character} scalar. Name of the argument passed.}
}
\value{
Stops the function if the file does not exist.
}
\description{
This is an internal function that is not meant to be used outside
of the package. It determines whether a specific file exists at the end of the
path given to the function.
}
\author{
Zheng Wei
}
