% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUtils.R
\name{checkPathExist}
\alias{checkPathExist}
\title{Check if path exists}
\usage{
checkPathExist(filePath, argname)
}
\arguments{
\item{filePath}{\code{Character} scalar. Path of interest.}

\item{argname}{\code{Character} scalar. Name of the argument passed.}
}
\value{
Stops the function if the path does not exist.
}
\description{
This is an internal function that is not meant to be used outside
of the package. It determines whether the path passed to the function exists.
}
\author{
Zheng Wei
}
