% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapterremoval.R
\name{identify_adapters}
\alias{identify_adapters}
\title{identify adapters for paired-end reads}
\usage{
identify_adapters(file1, file2, ..., interleaved = FALSE, overwrite = FALSE)
}
\arguments{
\item{file1}{\code{Character} vector. It can be file paths with #1 mates paired with file paths in file2
And it can also be interleaved file paths when argument interleaved=\code{TRUE}}

\item{file2}{\code{Character} vector. It contains file paths with #2 mates paired with file paths in seq1.
For interleaved paired-end sequencing files(argument interleaved=\code{TRUE}),it will be ignored.}

\item{...}{Additional arguments to be passed on to the binaries. See below for details.}

\item{interleaved}{\code{Logical}. Set \code{TRUE} when files are interleaved paired-end sequencing data.}

\item{overwrite}{\code{Logical}. Force overwriting of existing files if setting \code{TRUE}.}
}
\description{
This function can be use to call \code{AdapterRemoval} that wrapped in shared library for adapters identifying.
}
\details{
All additional arguments in ... are interpreted as additional parameters to be passed on to
bowtie2_build. All of them should be \code{Character} or \code{Numeric} scalar. You can put all aditional
arguments in one \code{Character}(e.g. "--threads 8") with white space splited just like command line,
or put them in different \code{Character}(e.g. "--threads","8").Note that some arguments to the
identify_adapters will be ignored if they are already handled as explicit function arguments. See the output of
\code{adapterremoval_usage()} for details about available parameters.
}
\references{
Schubert, Lindgreen, and Orlando (2016). AdapterRemoval v2: rapid
adapter trimming, identification, and read merging.
BMC Research Notes, 12;9(1):88.
}
