% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2_build_usage}
\alias{bowtie2_build_usage}
\title{Print available arguments that can be passed to bowtie2_build()}
\usage{
bowtie2_build_usage()
}
\value{
Information about available arguments that can be passed to bowtie2_build()
}
\description{
Calling bowtie2_build_usage() prints the available arguments that 
can be passed to the ... argument of the bowtie2_build() function of the package.  
Note that some arguments are invalid if they are already handled as explicit 
function arguments.
}
\examples{
bowtie2_build_usage()
}
\references{
Langmead B, Salzberg S.
Fast gapped-read alignment with Bowtie 2. Nature Methods. 2012, 9:357-359.
}
\author{
Zheng Wei
}
