% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2_usage}
\alias{bowtie2_usage}
\title{Print available arguments that can be passed to bowtie2()}
\usage{
bowtie2_usage()
}
\value{
Information about available arguments that can be passed to bowtie2().
}
\description{
Calling bowtie2_usage() prints the available arguments that can
be passed to the ... argument of the bowtie2() function of the package.  
Note that some arguments are invalid if they are already handled as explicit 
function arguments.
}
\examples{
bowtie2_usage()
}
\references{
Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read
alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
\author{
Zheng Wei
}
