% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2_version}
\alias{bowtie2_version}
\title{Print version information of bowtie2-2.4.4}
\usage{
bowtie2_version(indexType = "s")
}
\arguments{
\item{indexType}{\code{Character} scalar. Defaults to 's' which shows version
number of the small bowtie-align binary. 
Can optionally pass 'l' which shows version number of the large bowtie2-align binary.}
}
\value{
An invisible \code{Integer} of call status.
The value is 0 when there is not any mistakes
Otherwise the value is non-zero.
}
\description{
Print version information of bowtie2-2.4.4
}
\examples{
\donttest{
bowtie2_version()
bowtie2_verision(indexType = 'l')
}
}
\references{
Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read
alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
\author{
Zheng Wei, Rahul Varki
}
