% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2_version}
\alias{bowtie2_version}
\title{Print version information of bowtie2-2.4.4}
\usage{
bowtie2_version()
}
\value{
An invisible \code{Integer} of call status.
The value is 0 when there is not any mistakes
Otherwise the value is non-zero.
}
\description{
Calling bowtie2_version() prints the version information of 
the bowtie package used.
}
\examples{
bowtie2_version()
}
\references{
Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read
alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
\author{
Zheng Wei
}
